Handled form filling in AddSightingActivity and adding a bird sighting into the database.
With this commit, the AddSightingActivity should be in a working state. The location is updated automatically given the app has access and the GPS is turned on. The date and time field is also auto-filled to the current date and time. Validity of the form is checked by ensuring no field is left empty.